# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Builder platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface User {
  id: string;
  name: string;
  email: string;
  avatar?: string | null;
  onboardingCompleted: boolean;
  skillLevel?: string | null;
  category?: string | null;
  adSource?: string | null;
  country?: string | null;
  language?: string | null;
  credits: number;
  creditsUsed: number;
  plan: string;
  createdAt: string;
}

export interface RegisterBody {
  name: string;
  email: string;
  password: string;
  country?: string;
  language?: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface OnboardingBody {
  skillLevel: string;
  category: string;
  adSource: string;
}

export interface UpdateProfileBody {
  name?: string;
  avatar?: string | null;
}

export interface AvatarResponse {
  avatarUrl: string;
}

export interface Project {
  id: string;
  name: string;
  description?: string | null;
  url?: string | null;
  category?: string | null;
  userId: string;
  messageCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateProjectBody {
  description: string;
  url?: string | null;
  category?: string | null;
}

export interface UpdateProjectBody {
  name?: string;
  description?: string | null;
  url?: string | null;
}

export interface GenerateNameBody {
  description: string;
}

export interface GenerateNameResponse {
  name: string;
}

export type MessageRole = (typeof MessageRole)[keyof typeof MessageRole];

export const MessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface Message {
  id: string;
  projectId: string;
  role: MessageRole;
  content: string;
  thinkingSteps?: string[] | null;
  attachmentUrl?: string | null;
  createdAt: string;
}

export interface SendMessageBody {
  content: string;
  attachmentUrl?: string | null;
}

export interface UploadResponse {
  url: string;
}

export type UploadAvatarBody = {
  file?: Blob;
};

export type UploadFileBody = {
  file?: Blob;
};