# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Builder platform API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: user
    description: User profile operations
  - name: projects
    description: Project/chat operations
  - name: messages
    description: Message operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "200":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid input or user already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout user
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /user/onboarding:
    post:
      operationId: completeOnboarding
      tags: [user]
      summary: Complete user onboarding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingBody"
      responses:
        "200":
          description: Onboarding complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /user/profile:
    put:
      operationId: updateProfile
      tags: [user]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /user/avatar:
    post:
      operationId: uploadAvatar
      tags: [user]
      summary: Upload user avatar
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Avatar uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvatarResponse"

  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List user projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "200":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  /projects/generate-name:
    post:
      operationId: generateProjectName
      tags: [projects]
      summary: Generate a project name from description
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateNameBody"
      responses:
        "200":
          description: Generated name
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateNameResponse"

  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    put:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /projects/{projectId}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Messages list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message and get AI response
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "200":
          description: AI response message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  /projects/{projectId}/messages/stream:
    post:
      operationId: streamMessage
      tags: [messages]
      summary: Send a message and stream AI response
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "200":
          description: Streaming SSE response
          content:
            text/event-stream:
              schema:
                type: string

  /projects/{projectId}/upload:
    post:
      operationId: uploadFile
      tags: [projects]
      summary: Upload file for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: File uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]

    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        avatar:
          type: string
          nullable: true
        onboardingCompleted:
          type: boolean
        skillLevel:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        adSource:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        language:
          type: string
          nullable: true
        credits:
          type: number
        creditsUsed:
          type: number
        plan:
          type: string
        createdAt:
          type: string
      required: [id, name, email, onboardingCompleted, credits, creditsUsed, plan, createdAt]

    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        country:
          type: string
        language:
          type: string
      required: [name, email, password]

    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]

    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]

    OnboardingBody:
      type: object
      properties:
        skillLevel:
          type: string
        category:
          type: string
        adSource:
          type: string
      required: [skillLevel, category, adSource]

    UpdateProfileBody:
      type: object
      properties:
        name:
          type: string
        avatar:
          type: string
          nullable: true

    AvatarResponse:
      type: object
      properties:
        avatarUrl:
          type: string
      required: [avatarUrl]

    Project:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        url:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        userId:
          type: string
        messageCount:
          type: number
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, userId, messageCount, createdAt, updatedAt]

    CreateProjectBody:
      type: object
      properties:
        description:
          type: string
        url:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
      required: [description]

    UpdateProjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        url:
          type: string
          nullable: true

    GenerateNameBody:
      type: object
      properties:
        description:
          type: string
      required: [description]

    GenerateNameResponse:
      type: object
      properties:
        name:
          type: string
      required: [name]

    Message:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        thinkingSteps:
          type: array
          items:
            type: string
          nullable: true
        attachmentUrl:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, projectId, role, content, createdAt]

    SendMessageBody:
      type: object
      properties:
        content:
          type: string
        attachmentUrl:
          type: string
          nullable: true
      required: [content]

    UploadResponse:
      type: object
      properties:
        url:
          type: string
      required: [url]