# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Builder platform API
 * OpenAPI spec version: 0.1.0
 */
import type { MessageRole } from "./messageRole";

export interface Message {
  id: string;
  projectId: string;
  role: MessageRole;
  content: string;
  thinkingSteps?: string[] | null;
  attachmentUrl?: string | null;
  createdAt: string;
}