# lib/api-zod/src/generated/types/project.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Builder platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Project {
  id: string;
  name: string;
  description?: string | null;
  url?: string | null;
  category?: string | null;
  userId: string;
  messageCount: number;
  createdAt: string;
  updatedAt: string;
}