# lib/api-zod/src/generated/types/sendMessageBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Builder platform API
 * OpenAPI spec version: 0.1.0
 */

export interface SendMessageBody {
  content: string;
  attachmentUrl?: string | null;
}