# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Builder platform API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: string;
  name: string;
  email: string;
  avatar?: string | null;
  onboardingCompleted: boolean;
  skillLevel?: string | null;
  category?: string | null;
  adSource?: string | null;
  country?: string | null;
  language?: string | null;
  credits: number;
  creditsUsed: number;
  plan: string;
  createdAt: string;
}